backend/requirements.base.txt
# Heavy, rarely-changing dependencies
# These are cached in a separate Docker layer for optimal build performance

# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Configuration and environment
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# LangChain and AI (compatible versions) - HEAVY PACKAGES
langchain-core==0.3.15
langchain==0.3.7
langgraph==0.2.45
langgraph-checkpoint-postgres==2.0.6
psycopg[pool]==3.2.3

# Vector Database and Embeddings - VERY HEAVY PACKAGES
qdrant-client==1.16.1
sentence-transformers==5.1.2

# Scientific Computing - HEAVY PACKAGES (for A/B testing, ML models)
scipy==1.11.4
numpy==1.26.2

# LLM Provider SDKs
google-generativeai==0.8.3
